Snapnest/requirements.txt
asgiref==3.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==42.0.8
defusedxml==0.7.1
Django==6.0
django-cors-headers==4.9.0
django-debug-toolbar==6.1.0
django-extensions==3.2.3
django-sslserver==0.22
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
easy-thumbnails==2.10.1
git-filter-repo==2.47.0
idna==3.11
MarkupSafe==3.0.3
oauthlib==3.3.1
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==24.1.0
python-decouple==3.8
python3-openid==3.2.0
redis==7.1.0
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.4
urllib3==2.6.2
Werkzeug==3.0.2